cmd: Finish config move
Started the process in the previous commit, but forgot to
actually move all the relevant code over.
