+++ cmd/config.go
@@ -13,3 +13,30 @@
 // limitations under the License.
 
 package cmd
+
+import (
+	"fmt"
+
+	"github.com/spf13/viper"
+)
+
+var (
+	cfgFile string
+)
+
+// initConfig reads in config file and ENV variables if set.
+func initConfig() {
+	if cfgFile != "" { // enable ability to specify config file via flag
+		viper.SetConfigFile(cfgFile)
+	}
+
+	viper.SetEnvPrefix("hubble")
+	viper.SetConfigName(".hubble") // name of config file (without extension)
+	viper.AddConfigPath("$HOME")   // adding home directory as first search path
+	viper.AutomaticEnv()           // read in environment variables that match
+
+	// If a config file is found, read it in.
+	if err := viper.ReadInConfig(); err == nil {
+		fmt.Println("Using config file:", viper.ConfigFileUsed())
+	}
+}
